Move Spatial to it's own library due to lucene spatial depending on lgpl.
