+++ src/Examine.Lucene.Spatial/Search/ExamineLuceneCircle.cs
