+++ src/Examine.Lucene.Spatial/Search/ExamineLuceneEmptyShape.cs
