+++ src/Examine.Lucene.Spatial/Search/ExamineLucenePoint.cs
