+++ src/Examine.Lucene.Spatial/Search/ExamineLuceneRectangle.cs
