+++ src/Examine.Lucene.Spatial/Search/ExamineLuceneShape.cs
