+++ src/Examine.Lucene.Spatial/Search/ExamineLuceneShapeCollection.cs
