+++ src/Examine.Lucene.Spatial/Search/Spatial4nShapeFactory.cs
