+++ src/Examine.Lucene/Examine.Lucene.csproj
@@ -21,7 +21,6 @@
       <Version>4.8.0-beta00016</Version>
     </PackageReference>
     <PackageReference Include="Lucene.Net.Replicator" Version="4.8.0-beta00016" />
-    <PackageReference Include="Lucene.Net.Spatial" Version="4.8.0-beta00016" />
     <PackageReference Include="System.Threading">
       <Version>4.3.0</Version>
     </PackageReference>
